Extract todo index read model
