+++ examples/control_plane/todo-index-rollout-status-smoke.py
@@ -10,7 +10,8 @@ ROOT = Path(__file__).resolve().parents[2]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from loopx.status import build_todo_index
+from loopx import status as status_module
+from loopx.control_plane.todos.todo_index import build_todo_index as build_todo_index_read_model
 
 
 def write_event(path: Path, payload: dict[str, object]) -> None:
@@ -47,16 +48,55 @@ def main() -> None:
             },
         )
 
-        index = build_todo_index(
+        history = {"goals": [{"id": "loopx-meta"}]}
+        event_only_index = status_module.build_todo_index(
             queue={"items": []},
-            history={"goals": [{"id": "loopx-meta"}]},
+            history=history,
             runtime_root=runtime_root,
         )
+        queue = {
+            "items": [
+                {
+                    "goal_id": "loopx-meta",
+                    "agent_todos": {
+                        "items": [
+                            {
+                                "todo_id": "todo_done_status",
+                                "status": "open",
+                                "title": "Confirm rollout status merge",
+                                "text": "Confirm rollout status merge",
+                            }
+                        ]
+                    },
+                }
+            ]
+        }
+        index = status_module.build_todo_index(
+            queue=queue,
+            history=history,
+            runtime_root=runtime_root,
+        )
+        read_model_index = build_todo_index_read_model(
+            queue=queue,
+            history=history,
+            runtime_root=runtime_root,
+            public_safe_compact_text=status_module.public_safe_compact_text,
+        )
 
+    event_only_items = event_only_index["items"]
+    assert len(event_only_items) == 1, event_only_items
+    assert event_only_items[0]["source"] == "rollout_event_log", event_only_items[0]
+
+    assert read_model_index == index, (read_model_index, index)
+    assert index["current_projected_count"] == 1, index
+    assert index["rollout_event_count"] == 2, index
     items = index["items"]
     assert len(items) == 1, items
     item = items[0]
     assert item["todo_id"] == "todo_done_status", item
+    assert item["source"] == "attention_queue", item
+    assert item["event_count"] == 2, item
+    assert item["event_kinds"] == ["todo_add", "todo_complete"], item
     assert item["latest_event_kind"] == "todo_complete", item
     assert item["latest_event_status"] == "done", item
     assert item["status"] == "done", item
